Move all complex components to widgets/
