+++ labextension/src/components/ExperimentInput.tsx
@@ -15,9 +15,13 @@
  */
 
 import * as React from 'react';
-import { ISelectOption, IExperiment, NEW_EXPERIMENT } from './LeftPanelWidget';
 import { Input } from './Input';
 import { Select } from './Select';
+import {
+  ISelectOption,
+  IExperiment,
+  NEW_EXPERIMENT,
+} from '../widgets/LeftPanelWidget';
 
 const regex: string = '^[a-z]([-a-z0-9]*[a-z0-9])?$';
 const regexErrorMsg: string =
