+++ labextension/src/widgets/cell-metadata/CellMetadataContext.ts
