+++ labextension/src/widgets/cell-metadata/CellMetadataEditor.tsx
@@ -19,13 +19,13 @@ import { Notebook, NotebookPanel } from '@jupyterlab/notebook';
 import TagsUtils from '../../utils/TagsUtils';
 import { isCodeCellModel } from '@jupyterlab/cells';
 import CloseIcon from '@material-ui/icons/Close';
-import ColorUtils from './ColorUtils';
+import ColorUtils from '../../utils/ColorUtils';
 import { CellMetadataContext } from './CellMetadataContext';
 import { Button, IconButton } from '@material-ui/core';
 import { CellMetadataEditorDialog } from './CellMetadataEditorDialog';
-import { Input } from '../Input';
-import { Select } from '../Select';
-import { SelectMulti } from '../SelectMulti';
+import { Input } from '../../components/Input';
+import { Select } from '../../components/Select';
+import { SelectMulti } from '../../components/SelectMulti';
 
 const CELL_TYPES = [
   { value: 'imports', label: 'Imports' },
