+++ labextension/src/widgets/cell-metadata/InlineCellMetadata.tsx
